Jetpack Licensing: Reorganize files in preparation for the Licensing Portal
