+++ client/jetpack-cloud/sections/licensing-portal/index.ts
@@ -15,6 +15,6 @@ import * as controller from './controller';
 import './style.scss';
 
 export default function () {
-	page( `/:locale/licenses`, controller.jetpackLicensingContext, makeLayout, clientRender );
-	page( `/licenses`, controller.jetpackLicensingContext, makeLayout, clientRender );
+	page( `/:locale/licensing-portal`, controller.jetpackLicensingContext, makeLayout, clientRender );
+	page( `/licensing-portal`, controller.jetpackLicensingContext, makeLayout, clientRender );
 }
