+++ client/jetpack-cloud/sections/licensing-portal/inspect-license-form/style.scss
