+++ client/jetpack-cloud/sections/licensing-portal/inspect-license-result/style.scss
