+++ client/jetpack-cloud/sections/licensing-portal/inspect-licenses/index.tsx
@@ -7,8 +7,8 @@ import React from 'react';
  * Internal dependencies
  */
 import Main from 'calypso/components/main';
-import InspectLicenseForm from 'calypso/jetpack-cloud/sections/licensing/inspect-license-form';
-import InspectLicenseResult from 'calypso/jetpack-cloud/sections/licensing/inspect-license-result';
+import InspectLicenseForm from 'calypso/jetpack-cloud/sections/licensing-portal/inspect-license-form';
+import InspectLicenseResult from 'calypso/jetpack-cloud/sections/licensing-portal/inspect-license-result';
 
 /**
  * Style dependencies
