+++ client/jetpack-cloud/sections/licensing-portal/jpcom-footer/index.tsx
