+++ client/jetpack-cloud/sections/licensing-portal/jpcom-footer/style.scss
