+++ client/state/licensing-portal/reducer/index.js
