+++ client/state/licensing-portal/reducer/inspect-license.js
